+++ cypress/plugins/bitbucket.js
@@ -88,7 +88,7 @@ async function prepareTestBitBucketRepo({ lfs }) {
   // postfix a random string to avoid collisions
   const postfix = Math.random()
     .toString(32)
-    .substring(2);
+    .slice(2);
   const testRepoName = `${repo}-${Date.now()}-${postfix}`;
 
   console.log('Creating repository', testRepoName);