+++ packages/netlify-cms-backend-azure/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.1](https://github.com/netlify/netlify-cms/compare/netlify-cms-backend-azure@1.3.0...netlify-cms-backend-azure@1.3.1) (2022-04-13)
+
+**Note:** Version bump only for package netlify-cms-backend-azure
+
+
+
+
+
 # [1.3.0](https://github.com/netlify/netlify-cms/compare/netlify-cms-backend-azure@1.2.2...netlify-cms-backend-azure@1.3.0) (2021-10-18)